Traverse app and config parent directories when looking up modules
When running something from the build folder, it was really hard to set
up the paths properly and it would have to be done manually in most
cases. Traversing up from the binary makes it a lot easier. It also
takes preference over the config dir because there may be clashes
between dir names and the ones in build dir should be preferred.